requirements.txt
fastapi[standard]
uvicorn
scikit-learn
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
requests